edge-servers/radio/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
asyncio-mqtt==0.16.1

# Database
prisma==0.12.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
aiofiles==23.2.1

# Audio processing and radio
pydub==0.25.1
gTTS==2.4.0  # Google Text-to-Speech
pyttsx3==2.90  # Text-to-speech
soundfile==0.12.1
librosa==0.10.1
scipy==1.11.4

# Radio and broadcasting
pyserial==3.5
# radio==0.1.0  # Radio control library - commented out, not available
# rtlsdr==0.2.93  # Software-defined radio - commented out, not available

# FCC and EAS compliance
xmltodict==0.13.0
lxml==4.9.3

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AI and Analysis
google-generativeai==0.3.2

# Audio Processing (additional)
# gtts==2.4.0  # Already included above
# pyttsx3==2.90  # Already included above

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
